discount

+++ Data/EntityModels/Book.cs
@@ -13,7 +13,7 @@ namespace TheBookCave.Data.EntityModels
         public double Rating { get; set; }
         public double Price { get; set; }
         public string Genre { get; set; }
-        public int Discount { get; set; }
+        public double Discount { get; set; }
 
         public double DiscountPrice { get; set; }
         public int BoughtCopies { get; set; }
